refactor(feedback): split declared activity and feedback controller and service
refs: #1859
